Modify tests to account for parallel graph
